Refactor into multiple files.
