+++ index.html
@@ -4,5 +4,5 @@
 <h1>Walden Journal Extractor</h1>
 <div id="app"></div>
 <script>
-require('./dist/tsx/index.js');
+require('./dist/gui/renderer/index.js');
 </script>
