+++ src/gui/main/index.ts
@@ -1,0 +1,39 @@
+import * as path from 'path';
+import * as url from 'url';
+
+import 'source-map-support/register'
+import { app, BrowserWindow } from 'electron';
+
+import './rpc-server';
+
+let win: BrowserWindow | null = null;
+
+function createWindow() {
+    win = new BrowserWindow({ width: 800, height: 600 });
+
+    win.loadURL(url.format({
+        pathname: path.join(__dirname, '..', '..', '..', 'index.html'),
+        protocol: 'file:',
+        slashes: true,
+    }));
+
+    win.webContents.openDevTools();
+
+    win.on('closed', () => {
+        win = null;
+    });
+}
+
+app.on('ready', createWindow);
+
+app.on('window-all-closed', () => {
+    if (process.platform !== 'darwin') {
+        app.quit();
+    }
+});
+
+app.on('activate', () => {
+    if (win === null) {
+        createWindow();
+    }
+});
