+++ src/gui/renderer/rpc-client.ts
@@ -1,22 +1,6 @@
-import { h, Component, render } from 'preact';
 import { ipcRenderer, IpcMessageEvent } from 'electron';
 
-import { GuiRequest, GuiResponse } from '../gui';
-
-export interface AppProps {
-    msg: string;
-}
-
-export interface AppState {
-}
-
-export class App extends Component<AppProps, AppState> {
-    render(props: AppProps, state: AppState) {
-        return <p>{props.msg}</p>;
-    }
-}
-
-render(<App msg="Hello."></App>, document.getElementById('app'));
+import { GuiRequest, GuiResponse } from '../rpc';
 
 interface RequestPromiseCallbacks {
     resolve: (response: GuiResponse) => void;
@@ -25,7 +9,7 @@ interface RequestPromiseCallbacks {
 
 const requestMap: Map<number, RequestPromiseCallbacks> = new Map();
 
-function sendRequest(request: GuiRequest): Promise<GuiResponse> {
+export function sendRequest(request: GuiRequest): Promise<GuiResponse> {
     ipcRenderer.send('gui-request', request);
     return new Promise((resolve, reject) => {
         requestMap.set(request.id, { resolve, reject });
@@ -44,13 +28,4 @@ ipcRenderer.on('gui-response', (event: IpcMessageEvent, response: GuiResponse) =
     } else {
         callbacks.reject(new Error(response.error));
     }
-});
-
-sendRequest({
-    id: Math.random(),
-    ms: 1000,
-}).then((response) => {
-    console.log('callback promise returned!', response);
-}).catch(e => {
-    console.log('an error occurred', e.message);
 });
