+++ src/gui/rpc.ts
@@ -1,0 +1,9 @@
+export interface GuiRequest {
+    ms: number;
+    id: number;
+}
+
+export interface GuiResponse {
+    id: number;
+    error: string | null;
+}
